[#38] feat: 현재 시간 체크 헬퍼 함수 생성
- utils 폴더에서 관
- 현재 시간을 체크하는 함수
- set state 인자로 받음
